+++ src/View/Content.php
@@ -7,7 +7,7 @@ namespace App\View;
 final class Content
 {
     /**
-     * @param callable(mixed, array<string, mixed>): string|array<string> $contentExtractor
+     * @param callable(mixed, array<string, mixed>): (string|array<int, string>) $contentExtractor
      */
     public function __construct(
         public readonly mixed $contentExtractor,